extract gameboard
