+++ src/components/mod.rs
@@ -1,1 +1,2 @@
 pub mod character_page;
+pub mod gameboard;
